ref: checkpoint connector methods 3/n
